requirements.txt
discord
PyPDF2
psycopg2-binary